refactor: restructure routing and add onboarding page
- Moved routing logic to a new Router.tsx file for better organization.
- Introduced an Onboarding page with a welcoming message.
- Updated CTAButton to navigate to the new Onboarding page.
- Removed lazy loading for Index and NotFound from App.tsx, now handled in Router.tsx.
