+++ src/App.tsx
@@ -2,26 +2,12 @@ import { Toaster as Sonner } from "@/components/ui/sonner";
 import { TooltipProvider } from "@/components/ui/tooltip";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import { RootLayout } from '@/components/RootLayout';
-import { lazy, Suspense } from 'react';
+import { Suspense } from 'react';
 import { GlobalLoadingOverlay } from "@/components/ui/loader-global-state";
+import { RouterProvider } from 'react-router-dom';
+import { router } from './Router.tsx';
 
 const queryClient = new QueryClient();
-
-const Index = lazy(() => import('@/pages/Index'));
-const NotFound = lazy(() => import('@/pages/NotFound'));
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <RootLayout />,
-    children: [
-      { index: true, element: <Index /> },
-      { path: '*', element: <NotFound /> },
-    ],
-  },
-]);
 
 const App = () => (
   <QueryClientProvider client={queryClient}>
