+++ src/components/ui/ctabutton.tsx
@@ -1,9 +1,11 @@
 import styled from 'styled-components';
+import { useNavigate } from 'react-router-dom';
 
 const CTAButton = () => {
+  const navigate = useNavigate();
   return (
     <StyledWrapper>
-      <button className="animated-button">
+      <button className="animated-button" onClick={() => navigate('/onboarding')}>
         <svg viewBox="0 0 24 24" className="arr-2" xmlns="http://www.w3.org/2000/svg">
           <path d="M16.1716 10.9999L10.8076 5.63589L12.2218 4.22168L20 11.9999L12.2218 19.778L10.8076 18.3638L16.1716 12.9999H4V10.9999H16.1716Z" />
         </svg>
