+++ src/pages/Onboarding.tsx
@@ -1,0 +1,11 @@
+
+const Onboarding = () => {
+  return (
+    <div className="min-h-screen flex flex-col items-center justify-center bg-slate-900 text-white">
+      <h1 className="text-4xl font-bold mb-4">Bienvenido a Open Mus</h1>
+      <p className="text-lg text-slate-300">¡Comienza tu experiencia de Mus online!</p>
+    </div>
+  );
+};
+
+export default Onboarding;
